southbound implementation has been improved to make it more general
