+++ southbound/common.go
@@ -1,0 +1,31 @@
+// Copyright 2019-present Open Networking Foundation
+//
+// Licensed under the Apache License, Configuration 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package southbound
+
+import (
+	"context"
+)
+
+// Key struct - can be extended, for now is ip:port
+type Key struct {
+	Key string
+}
+
+// Target common struct for connecting to gNMI/gNOI targets
+type Target struct {
+	Ctxt context.Context
+}
+
+//var targets = make(map[Key]interface{})
